Continued with adding intermediate nodes for Replicate and Chargestate.

+++ src/openms/source/FORMAT/CachedMzML.cpp
@@ -109,7 +109,7 @@ namespace OpenMS
   MSChromatogram CachedmzML::getChromatogram(Size id)
   {
     OPENMS_PRECONDITION(id >= 0, "Id needs to be larger than zero");
-    OPENMS_PRECONDITION(id < (int)getNrChromatograms(), "Id cannot be larger than number of chromatograms");
+    OPENMS_PRECONDITION(id < getNrChromatograms(), "Id cannot be larger than number of chromatograms");
 
     if ( !ifs_.seekg(chrom_index_[id]) )
     {
